+++ agent_bench/judging/__init__.py
@@ -1,0 +1,3 @@
+from agent_bench.judging.reasoning import build_reasoning_benchmark, reasoning_enabled
+
+__all__ = ["build_reasoning_benchmark", "reasoning_enabled"]